ANDROID: input: Add keyreset driver.
Add a platform device in the board file to specify a reset key-combo.
The first time the key-combo is detected a work function that syncs
the filesystems is scheduled. If all the keys are released and then
pressed again, it calls panic. Reboot on panic should be set for
this to work.

Includes subsequent work and fixes:

  Daniel Rosenberg <drosen@google.com>:
  input: Changed keyreset to act as a wrapper for keycombo.
  Daniel Rosenberg <drosen@google.com>:
  input: Made keyreset more robust
  Amit Pundir <amit.pundir@linaro.org>:
  android: skip building drivers as modules

Signed-off-by: Arve Hjønnevåg <arve@android.com>
Signed-off-by: Daniel Rosenberg <drosen@google.com>
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>

Bug: None
Patchset: input-keycombo

Signed-off-by: Mattias Nissler <mnissler@google.com>
Change-Id: If46ac8c478ad52af13c79522438201afbf9af59f