ANDROID: input: keychord: Add keychord driver
This driver allows userspace to receive notification when client
specified key combinations are pressed.
The client opens /dev/keychord and writes a list of keychords
for the driver to monitor.
The client then reads or polls /dev/keychord for notifications.
A client specified ID for the keychord is returned from read()
when a keychord press is detected.

Includes subsequent work and fixes:

  JP Abgrall <jpa@google.com>:
  input: misc: keychord: log when keychord triggered
  Colin Cross <ccross@android.com>:
  input: misc: keychord: move header to uapi

Signed-off-by: Mike Lockwood <lockwood@android.com>
Signed-off-by: JP Abgrall <jpa@google.com>
Signed-off-by: Colin Cross <ccross@android.com>

Bug: None
Patchset: input-keychord

Signed-off-by: Mattias Nissler <mnissler@google.com>
Change-Id: Ief6b69833505b705919e85c80df48396b145a1fe