ANDROID: panic: Add board ID to panic output
At times, it is necessary for boards to provide some additional information
as part of panic logs. Provide information on the board hardware as part
of panic logs.

It is safer to print this information at the very end in case something
bad happens as part of the information retrieval itself.

To use this, set global mach_panic_string to an appropriate string in the
board file.

Signed-off-by: Nishanth Menon <nm@ti.com>

Bug: None
Patchset: panic-board-id

Signed-off-by: Mattias Nissler <mnissler@google.com>
Change-Id: I1eaaf6a67a8b58f86bbf66ea6a0091ced64fcb88