ANDROID: bridge: Have tx_bytes count headers like rx_bytes.
Since rx_bytes accounting does not include Ethernet Headers in
br_input.c, excluding ETH_HLEN on the transmit path for consistent
measurement of packet length on both the Tx and Rx chains.

The clean way would be for Rx to include the eth header, but the
skb len has already been adjusted by the time the br code sees the skb.
This is only a temporary workaround until we can completely ignore or
cleanly fix the skb->len handling.

Signed-off-by: Ashish Sharma <ashishsharma@google.com>

Bug: None
Patchset: net-bridge-tx-byte-counting

Signed-off-by: Mattias Nissler <mnissler@google.com>
Change-Id: I511c85e54ea8cf79052533b60aa387d7d3a644ea