ANDROID: ARM: decompressor: Flush tlb before swiching domain 0 to client mode
If the bootloader used a page table that is incompatible with domain 0
in client mode, and boots with the mmu on, then swithing domain 0 to
client mode causes a fault if we don't flush the tlb after updating
the page table pointer.

v2: Add ISB before loading dacr.

Signed-off-by: Arve Hjønnevåg <arve@android.com>

Bug: None
Patchset: arm-decompressor-tlb-flush

Signed-off-by: Mattias Nissler <mnissler@google.com>
Change-Id: Ie7b8028ae9acdc7632dcdc2a02221394c1741b5a