ANDROID: hardlockup: detect hard lockups without NMIs using secondary cpus
Emulate NMIs on systems where they are not available by using timer
interrupts on other cpus.  Each cpu will use its softlockup hrtimer
to check that the next cpu is processing hrtimer interrupts by
verifying that a counter is increasing.

This patch is useful on systems where the hardlockup detector is not
available due to a lack of NMIs, for example most ARM SoCs.
Without this patch any cpu stuck with interrupts disabled can
cause a hardware watchdog reset with no debugging information,
but with this patch the kernel can detect the lockup and panic,
which can result in useful debugging info.

Includes subsequent fix:

  Brian Norris <briannorris@google.com>:
  ANDROID: kernel/watchdog: fix unused variable warning

Signed-off-by: Colin Cross <ccross@android.com>
Signed-off-by: Brian Norris <briannorris@google.com>

Bug: None
Patchset: watchdog-secondary-cpus

Signed-off-by: Mattias Nissler <mnissler@google.com>
Change-Id: I5e0e8fb2b760fa254e397a5a0c342a242860601d