ANDROID: tcp: add a sysctl to config the tcp_default_init_rwnd
The default initial rwnd is hardcoded to 10.

Now we allow it to be controlled via
  /proc/sys/net/ipv4/tcp_default_init_rwnd
which limits the values from 3 to 100

This is somewhat needed because ipv6 routes are
autoconfigured by the kernel.

See "An Argument for Increasing TCP's Initial Congestion Window"
in https://developers.google.com/speed/articles/tcp_initcwnd_paper.pdf

Includes subsequent fix:

  Dmitry Shmidt <dimitrysh@google.com>
  tcp: fix tcp_default_init_rwnd() for 4.1

Signed-off-by: JP Abgrall <jpa@google.com>
Signed-off-by: Dmitry Shmidt <dimitrysh@google.com>

Bug: 12020135
Patchset: tcp-rwnd-sysctl

Signed-off-by: Mattias Nissler <mnissler@google.com>
Change-Id: I09311769d253793252fc655158e765de6f6de89e