ANDROID: Shrink ashmem directly through shmem_fallocate
When ashmem_shrink is called from direct reclaim on a user thread, a
call to do_fallocate will check for permissions against the security
policy of that user thread.  It can thus fail by chance if called on a
thread that isn't permitted to modify the relevant ashmem areas.

Because we know that we have a shmem file underneath, call the shmem
implementation of fallocate directly instead of going through the
user-space interface for fallocate.

FIX=DMS06243560
Area: Kernel/Linux Kernel

Signed-off-by: Jeff Vander Stoep <jeffv@google.com>

Bug: 21951515
Patchset: staging-android-ashmem

Signed-off-by: Mattias Nissler <mnissler@google.com>
Change-Id: I88d783e24230d00389b0c2768324688cd4ea6212