UPSTREAM: dm: don't save and restore bi_private
Device mapper used the field bi_private to point to dm_target_io. However,
since kernel 3.15, the bi_private field is unused, and so the targets do
not need to save and restore this field.

This patch removes code that saves and restores bi_private from dm-cache,
dm-snapshot and dm-verity.

Signed-off-by: Mikulas Patocka <mpatocka@redhat.com>
Signed-off-by: Mike Snitzer <snitzer@redhat.com>

Bug: None
Patchset: dm-verity

Signed-off-by: Mattias Nissler <mnissler@google.com>
Change-Id: I918dae5fd4c1fd166efa80ccd8ffa7803216f2b3