dm: export a table+mapped device to the ioctl interface
If a mapped device and table is configured without traversing the dm-ioctl
interface (dm-fs-style), then it will not be bound to a name or uuid. This
means that it will be inaccessible for userspace management and udev
will be unhappy with the lack of a name or uuid.

The function added in this change performs the required association to
transition to being managed by the ioctl interface.

Signed-off-by: Will Drewry <wad@chromium.org>
Signed-off-by: Kees Cook <keescook@chromium.org>

Bug: 27242350

(git am from https://patchwork.kernel.org/patch/8367701/)
Signed-off-by: Kees Cook <keescook@google.com>
Change-Id: I8ca28c24acaf8557332b0bec62dedf6001a79a9c