FROMLIST: efi: Introduce EFI bootloader control driver
This driver intercepts system reboot requests and populates the
LoaderEntryOneShot EFI variable with the user-supplied reboot
argument. EFI bootloaders such as Gummiboot will consume this
variable and use it to control which OS is booted next.

We use this with Android where reboot() tells the kernel that
we want to boot into recovery or other non-default OS environment.

It is the bootloader's job to guard against this variable being
uninitialzed or containing invalid data, and just boot normally
if that is the case.

Signed-off-by: Matt Gumbel <matthew.k.gumbel@intel.com>
Signed-off-by: Mohamed Abbas <mohamed.abbas@intel.com>
Signed-off-by: Constantin Musca <constantin.musca@intel.com>
Signed-off-by: Daniel Baluta <daniel.baluta@intel.com>

BUG=27540453

(git am from https://patchwork.kernel.org/patch/8617801/)
Signed-off-by: Stefan Stanacar <stefan.stanacar@intel.com>
Change-Id: I4c6c714b1a67ab12a01aa0e5cfe27cb00e65294d