ANDROID: wlcore: Disable filtering in AP role
When you configure (set it up) a STA interface, the driver
install a multicast filter. This is normal behavior, when
one application subscribe to multicast address the filter
is updated. When Access Point interface is configured, there
is no filter installation and the "filter update" path is
disabled in the driver.

The problem happens when you switch an interface from STA
type to AP type. The filter is installed but there are no
means to update it.

Signed-off-by: Dmitry Shmidt <dimitrysh@google.com>

Bug: None
Patchset: wlcore-disable-filtering

Signed-off-by: Mattias Nissler <mnissler@google.com>
Change-Id: I1853c2e2c592216e0e88fb4f33e46bf094a4aa4a